Simplify the server's logic
The Python script doesn't print the downloading messages anymore.
We now check on the server if the classification model directory exists beforehand.
If it doesn't, we log the information while the
classification processes.
We therefore have to call the classification
script only once.
The model directory is now customizable with the `SM_MODEL_DIR`
enrivonment variable.